Separate media converters from video converters in another file, fix an issue where output formats werent set, and add video info converters.
